+++ src/Ema/Multi/Generic/Example.hs
@@ -8,7 +8,11 @@ import Data.Generics.Sum.Any (AsAny (_As))
 import Data.SOP (I (..), NP (..))
 import Ema.App qualified as Ema
 import Ema.Asset qualified as Asset
-import Ema.Multi.Generic
+import Ema.Multi.Generic (WithConstModel (..), WithModel (..))
+import Ema.Multi.Generic.Motley (
+  MotleyModel (..),
+  MotleyRoute (MotleyRouteSubRoutes),
+ )
 import Ema.Route.Class (IsRoute (..))
 import Ema.Route.Encoder
 import Ema.Route.Extra
